# swagger.yml
openapi: 3.0.3
info:
  title: 出席管理アプリケーション
  description: |-
    みやは変態です。
  version: 1.0.0
servers:
  - url: https://n3a.miyayu.xyz/api/
...
paths:
  /student:
    post:
      tags:
        - Student
      summary: 生徒を登録します
      description: 生徒を新しく登録します
      operationId: addStudent
      parameters:
        - name: studentId
          in: query
          description: 学籍番号
          required: true
          schema:
            type: string
        - name: name
          in: query
          description: 生徒名
          required: true
          schema:
            type: string
        - name: departmentId
          description: 科目ID
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 生徒の登録に成功したとき
        '400':
          description: 入力値が不正な時
    get:
      tags:
        - Student
      summary: 生徒の一覧を取得します
      description: 登録されているすべての生徒を取得します
      operationId: getAllStunet
      responses:
        '200':
          description: 生徒一覧が出力されます
          content:
            application/json:
              schema:
                type: object
                properties:
                  students:
                    type: array
                    items:
                      $ref: '#/components/schemas/Student'
  /stundent/{studentAutoId}:
    get:
      tags:
        - Student
      summary: 生徒をIDから取得します
      description: 一致する生徒を1名のみ返します。
      operationId: getStudentById
      parameters:
        - name: studentAutoId
          in: path
          description: 学籍番号
          required: true
          example: 10
          schema:
            type: string
      responses:
        '200':
          description: 生徒を返す
          content:
            application/json:
              schema:
                type: object
                required:
                  - student
                properties:
                  student:
                    nullable: true
                    $ref: '#/components/schemas/Student'

        '400':
          description: 不正な学籍番号のとき
...
components:
  schemas:
...
    Student:
      type: object
      required:
        - id
        - studentId
        - name
        - departmentId
        - icId
      properties:
        id:
          type: integer
          description: システム上で管理されているID
          example: 3
        studentId:
          type: string
          description: 学生ID
          example: G021C1000
        name:
          type: string
          description: 氏名
          example: 鈴木 鈴木
        departmentId:
          type: number
          description: 学科ID
          example: 1
        icId:
          type: string
          description: ICカードのIDm（固有識別子）
          example: 0xabcdef
          nullable: true
...